split out check-lock-file to its own file
